+++ agent/src/main/kotlin/com/jetbrains/teamcity/plugins/unrealengine/agent/buildgraph/DistributedBuildSettingsCreator.kt
@@ -59,6 +59,7 @@ class DistributedBuildSettingsCreator(
                     runnerInternalSettings.compositeBuildId,
                 )
             }
+
             is BuildGraphRunnerInternalSettings.RegularBuildSettings -> {
                 DistributedBuildSettings.RegularBuildSettings(
                     sharedStorageDir,